Try to avoid segfaults
Do this by removing the dependency on 'trap' and 'on_exit' to avoid
segfaults in the ruby interpreter.  Also deal better with additional
interrupts raised in the rollback code, particularly SIGTERM, which
was not stubbed out in the past.

This code is now entirely dependent on standard Ruby exception
handling to deal with interrupts.

Signed-off-by: Daniel Farina <daniel@heroku.com>